Вернула в отдельный файл types
