+++ src/components/list-offers/list-offers.tsx
@@ -1,4 +1,4 @@
-import { Offer } from '../../mocks/offers';
+import type { Offer } from '../../types/types';
 import CardOffer from '../card/card';
 import { useState } from 'react';
 
@@ -9,7 +9,9 @@ type ListOffersProps = {
 
 
 function ListOffers ({offers}: ListOffersProps): JSX.Element {
+
   const [activeCardId, setActiveCardId] = useState<string | null>(null);
+
   return(
     <div className="cities__places-list places__list tabs__content">
       {offers.map((offer) => (
