+++ src/mocks/offers.ts
@@ -1,47 +1,5 @@
-type OfferHost = {
-  id: number;
-  name: string;
-  isPro: boolean;
-  avatarUrl: string;
-}
+import { Offers } from '../types/types';
 
-type OfferLocation = {
-  latitude:number;
-  longitude: number;
-  zoom: number;
-}
-
-type OfferCityLocation = {
-  latitude: number;
-  longitude: number;
-  zoom: number;
-}
-
-type OfferCity = {
-  name: string;
-  location :OfferCityLocation;
-}
-
-type Offer = {
-city: OfferCity;
-previewImage: string;
-images: string[];
-title: string;
-isFavorite:boolean;
-isPremium:boolean;
-rating: number;
-type: string;
-bedrooms: number;
-maxAdults: number;
-price:number;
-goods: string[];
-host: OfferHost;
-description:string;
-location: OfferLocation;
-id:string;
-}
-
-type Offers = Offer[]
 
 const offersMock: Offers = [
   {
@@ -4128,5 +4086,5 @@ const offersMock: Offers = [
 
 
 export {offersMock};
-export type {Offers, Offer};
+
